web_app/requirements_web.txt
streamlit>=1.28.0
PyPDF2>=3.0.0
python-docx>=1.1.0
nltk>=3.8.0
textblob>=0.17.0
wordcloud>=1.9.0
matplotlib>=3.7.0
pandas>=2.0.0
plotly>=5.15.0
pillow>=10.0.0
io
base64